requirements-dev.txt
# Core FastAPI dependencies (Python >=3.10 compatible)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
pillow>=10.4.0

# ML/AI dependencies (Python >=3.10 compatible)
transformers>=4.40.0
torch>=2.4.0
scikit-learn>=1.5.0
nltk>=3.9.0
sentence-transformers>=3.0.0
numpy>=2.0.0
pandas>=2.2.0

# HTTP and utility dependencies (Python >=3.10 compatible)
httpx>=0.27.0
requests>=2.32.0
pyyaml>=6.0.2
loguru>=0.7.2
tenacity>=9.0.0
beautifulsoup4>=4.12.0

# Testing dependencies (Python >=3.10 compatible)
pytest>=8.3.0
pytest-asyncio>=0.24.0

# Development dependencies (Python >=3.10 compatible)
black>=24.8.0
flake8>=7.1.0
